drivers: octeontx2: memtest: Retry on failing SPI lock
commit 679cccdccb2c1321bda92597b8bbe28ecb887634 from
git@git.assembla.com:cavium/WindRiver.linux.git

Due to concurrent access to SPI bus it may happen that SMC call
fails. Handle this case and retry up to 50 times with a short delay.

In local test it happened once in 1000 boots that 37 retries where
needed, so use 50 as a maximum.

Change-Id: I036265bd7a8b16ba1092445cdc96a9e43157ae15
Signed-off-by: Michael Lawnick <michael.lawnick@nokia.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/108590
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>