net: phy: nxp-c45-tja11xx: handle FUSA irq
commit 0eb90f7efd3d2dbf49742407351758dcd728508d from
https://github.com/nxp-auto-linux/linux

TJA1120 and TJA1103 have a set of functional safety hardware tests
executed after every reset, and when the tests are done, the IRQ line is
asserted. For the moment, the purpose of these handlers is to acknowledge
the IRQ and not to check the FUSA tests status.

Issue: ALB-10440

Signed-off-by: Radu Pirea (NXP OSS) <radu-nicolae.pirea@oss.nxp.com>
Reviewed-by: Andrew Lunn <andrew@lunn.ch>
Link: https://lore.kernel.org/r/20230731091619.77961-9-radu-nicolae.pirea@oss.nxp.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 425c8348df7b868bc25dce85758e5dc3d63de91d)
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>