arm64: dts: ti: k3-am69-sk-csi2-ov5640: Fix powerdown GPIO
commit 564a1e8f3625780e63bbf1983578c95ebd2ec8e4 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

While adding the overlay for OV5640 with AM69 SK, the powerdown
GPIO lines were modelled as gpio-hog which was pulled low by default.
While this results in a working configuration normally, when runtime
PM is active, the sensor resume sequence is affected which resulted in
stream failures. Model the power-down GPIOs properly so that it can be
controlled from OV5640 driver properly.

Fixes: 2441d6174947 ("arm64: dts: ti: Add DT overlay for OV5640 sensor")

Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>