HSD #16020988342: taprio: Add support for frame preemption offload
commit 3ea5cc5c880d0de3dc74dbe256fd1ab05ba4e3a4 from
https://github.com/altera-opensource/linux-socfpga.git socfpga-6.1.55-lts.

Adds a way to configure which traffic classes are marked as
preemptible and which are marked as express.

Even if frame preemption is not a "real" offload, because it can't be
executed purely in software, having this information near where the
mapping of traffic classes to queues is specified, makes it,
hopefully, easier to use.

taprio will receive the information of which traffic classes are
marked as express/preemptible, and when offloading frame preemption to
the driver will convert the information, so the driver receives which
queues are marked as express/preemptible.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>
Signed-off-by: Malli C <mallikarjuna.chilakala@intel.com>
Signed-off-by: Muhammad Husaini Zulkifli <muhammad.husaini.zulkifli@intel.com>
Signed-off-by: Rohan G Thomas <rohan.g.thomas@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>