phy: lynx-28g: add support for backplane modes through PHY_MODE_ETHERNET_LINKMODE
commit c9c9d09e10215eeba42581a8b1a7d278205dab69 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

The Lynx 28G SerDes lanes allow the tuning of their signal equalization
parameters as required by the IEEE 802.3 clauses for copper backplanes.

These modes are selected when there is a backplane AN/LT block in charge,
and it uses phy_set_mode_ext(PHY_MODE_ETHERNET_LINKMODE).

Only the single-lane link modes (1000Base-KX, 10GBase-KR, 25GBase-KR)
are supported here (40GBase-KR, 100GBase-KR - not yet). And only the
initial configuration for these link modes is present, not the dynamic
equalization changes. That will come as a separate change.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>