net: ethernet: ti: j721e-cpsw-virt-mac: Exit notification handler if ndev is NULL
The notification received from the core networking subsystem should not
provide a NULL pointer for the net_device. It is observed that when a
virtual networking device is created, its associated notifier chain
results in a NULL pointer being handed over to the
virt_cpsw_inetaddr_event() notification handler.

Exit the virt_cpsw_intetaddr_event() handler in such a situation.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
[Xulin: picked from ti-processor-sdk-linux-adas-j784s4-evm-09_00_01_02]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>