firmware: xilinx: remove clock_setrate and clock_getrate api
commit eb2da12bdff26fde55d9c23392e653380172e41c from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

As per the current code base, PM_CLOCK_SETRATE and PM_CLOCK_GETRATE
APIs are not supported for the runtime operations. In the case of
ZynqMP returning an error from TF-A when there is any request to
access these APIs and for Versal also it is returning an error like
NO_ACCESS from the firmware. So, just removing the unused code to
avoid the confusion around these APIs.

Also, there is no issue with the backward compatibility as these APIs
were never used since implemented. Hence no need to bump up the
version of the feature check API as well.

Signed-off-by: Ronak Jain <ronak.jain@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>