MA-21538-2 clk: imx: clk-imx8qxp: Parent should be initialized earlier than the clock
commit 0831bdd5efd82d77f01b46e90d677f5d7b3810d4 from
https://github.com/nxp-imx/linux-imx.git

The initialization order of SCU clocks affects the sequence of SCU clock
resume. If there are no other effects, the earlier the initialization,
the earlier the resume. During SCU clock resume, the clock rate is
restored. As SCFW guidelines, it is imperative to configure the parent
clock rate before configuring the child rate.

Change-Id: Ib475396c86fd8e2f8ee5107ee681dc24fabefb93
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>