llce: Apply W/A for generic_handle_irq with RT_FULL
commit d93c127ec819c5fa5f22345ab111e27fc5aca18e from
https://github.com/nxp-auto-linux/linux

According to Documentation/gpio/driver.txt, "W/A: The
generic_handle_irq() is expected to be called with IRQ disabled, so the
IRQ core will complain if it is called from an IRQ handler which is
forced to a thread. The "fake?" raw lock can be used to W/A this
problem".

Issue: ALB-10318
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>