llce: Apply W/A for generic_handle_irq with RT_FULL
commit dcaa44fade015848e357497d483ab7eaa9f57645 from
https://github.com/nxp-auto-linux/linux

According to Documentation/gpio/driver.txt, "W/A: The
generic_handle_irq() is expected to be called with IRQ disabled, so the
IRQ core will complain if it is called from an IRQ handler which is
forced to a thread. The "fake?" raw lock can be used to W/A this
problem".

Issue: ALB-10318
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>