linflex: Refactor interrupt flow
commit 124e5c241bac5dc24d6092f96282044f7350c818 from
https://github.com/nxp-auto-linux/linux

The current implementation is a mix between PIO and
polling. When characters are sent, the driver writes characters one
by one and waits for them to be sent. This is done while port lock
is acquired. RX interrupt handler is not called in time and
overflow occurs.
The proposed solution is to not transfer all characters in start_tx
and let the TX interrupt handler consume them.

From the user point of view, transferring a file will take the same
amount of time, in both implementations. The improvement can be seen
if we pay attention to the time spent in kernel. For example, the
time necessary to run 'cat' on 1M file:
- old approach:
real    1m32.188s
user    0m0.000s
sys     1m32.187s

- new approach:
real    1m32.547s
user    0m0.000s
sys     0m0.023s

Issue: ALB-10436
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
[fix: use spin_lock/unlock_*() functions to replace
uart_port_lock/unlock_*() functions]
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>