linflex: Wait for one more character in RX handler
commit 36d37ee1699b90389ad6e460210d9c91c575df73 from
https://github.com/nxp-auto-linux/linux

Linflex driver lose characters during copy-paste. When interrupts are
activated, Linflex can work in buffer mode only. According to the
manual, "If you do not know in advance how many bytes are to be
received, RDFL must not be programmed in advance. The reset value of
RDFL is 0. This ensures that the reception happens byte by byte."
Taking into account the buffer size, the buffer can easily overflow if
the character is not consumed in time (UARTSR.RMB is not cleared when a
new character is received), and this is happening in the "copy-paste"
scenario.
It was observed that, if the character is received just before exiting
the interrupt handler, Linux driver can't handle it in time and overflow
occurs. To avoid this, wait one more character in the interrupt before
exiting.
Taking into account the buffer size and the lack of flow control,
overflow is still possible, but with this patch, it is reduced
significantly.

Issue: ALB-10436
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>