drivers: mpam: Drop the improper assertions
The mpam_reprogram_ris_partid() and mpam_reset_ris() may be invoked on a
specific CPU by smp_call_on_cpu(). According to the implementation of
smp_call_on_cpu(), that also means these functions would be executed
in the worker kthread context. Both the
lockdep_assert_preemption_disabled() and lockdep_assert_held()
assertions are not suitable in this context because:
 - the lockdep_assert_preemption_disabled() will trigger false positive
   since the preempt is enabled in the worker kthread, but it will not
   migrate to other CPU since the worker kthread is bound to one CPU.
 - the implementation of lockdep_assert_held() is to check the lock
   acquired by the current process. So in a worker kthread, it
   definitely not acquire the &msc->lock.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>