watchdog: xilinx_wwdt: Add "xlnx,versal-wwdt-1.0" compatible
commit 6f5718a53545c67d0b60abe496d2e52e6803bbd0 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Add xlnx,versal-wwdt-1.0 compatible in xilinx_wwdt driver for backward
dts compatibility.

In the current xilinx_wwdt driver, watchdog is configured as
watchdog_set_nowayout with that stop call is not allowed by framework
itself, hence stop, set_timeout and set_pretimeout features are not
supported.

If driver is probed with xlnx,versal-wwdt-1.0 deprecated compatible, an
warning message is displayed and is recommended to switch to new bindings.
This is temporary workaround and will be removed in next release.

Signed-off-by: Srinivas Goud <srinivas.goud@amd.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>