arm64: dts: ti: k3-j784s4-evm-virt-mac-client: Fix Serdes Lane Mapping
commit ecfc9f11bc175c5fbbb2a17d664659ecfe4fcaba from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The serdes_ln_ctrl device-tree node in the overlay doesn't have mux
values for the Serdes 2 Lanes 0 and 1, due to which the mux values for
Serdes 2 Lanes 2 and 3 are assigned to Serdes 2 Lanes 0 and 1, while
Serdes 2 Lanes 2 and 3 are left unconfigured at their default values.

Fix this.

Fixes: ff7b39fd7004 ("arm64: dts: ti: k3-j784s4: Add overlay to enable CPSW9G Virt MAC Client")
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Ravi Gunasekaran <r-gunasekaran@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>