net: mii: add C73 link mode and base page helpers
commit a87111adf5168aa716a3a8fa0800a6bd2ac93ee5 from
https://github.com/nxp-imx/linux-imx.git

IEEE 802.3 clause 73 defines auto-negotiation for backplanes and copper
cable assemblies. This is a medium type (link mode) just like Ethernet
over twisted pairs (BASE-T / BASE-TX) and over two wires for automotive
(BASE-T1) for which the PHY library currently contains support.

As a minimal framework for backplane PHY drivers, introduce a set of
helpers that parse and interpret the base pages that are exchanged by
PHYs during the clause 73 negotiation.

The placement in the "legacy" mii code is perhaps not the best, but I
tried to put them somewhere accessible by phylib, phylink and non-phylib
drivers. Note that phylink also has its own phylink_resolve_c73() which
is more or less similar in purpose, but:

- it requires constructing a struct phylink_link_state which is deeply
  embedded with the phylink API and that may not be desirable for
  drivers

- the presence of some link modes in phylink's own
  phylink_c73_priority_resolution[] is "interesting", like
  ETHTOOL_LINK_MODE_2500baseX_Full_BIT, which is not a backplane mode
  negotiable through C73 at all. That comes from the xpcs driver which
  may have a non-standard C73 autoneg, and this makes it difficult for
  me to e.g. refactor phylink_resolve_c73() to use the more generic
  linkmode_c73_priority_resolution(). Also see the attached link where I
  had previously pointed this out.

Link: https://lore.kernel.org/netdev/20230516090009.ssq3uedjl53kzsjr@skbuf/
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>