LF-9952 clocksource: imx-tpm: Wait for CnV write to take effect
commit 8a37f2aba59cdc5ab5ad38bfe330c2ef58672ac9 from
https://github.com/nxp-imx/linux-imx.git

The value write into the TPM CnV can only be updated into the HW when
CNT increase. Additional writes to the CnV write buffer are ignored until
the register has been updated. So we need to check if the CnV has been
updated before continue. Wait for 1 CNT cycle in worst case.

Additionally, current return check is not correct, if a max_delta need
be set, it will return '-ETIME' wrongly due to the 'int' type cast,
so refine the check logic to fix it.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>