spi: spi-fsl-lpspi: Fix transmission when use CONT
commit c48dfb927834391852183fd696919ad5248edbf7 from
https://github.com/nxp-auto-linux/linux

Commit <6a130448498c> ("spi: lpspi: Fix wrong
transmission when don't use CONT") broken the transmission when CONT is
used.
TDIE interrupt should not be disabled in all cases. If CONT is used, and
TX transfer is not completed yet, but the isr handler is called because
there are characters to be received, TDIE is replaced with FCIE. When
the transfer is completed, SR_TDF is set but the isr isn't called again.

Issue: ALB-10320
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>