octeontx2-pf: Fix updating PFC configuration during running traffic
commit 4d90daa13f96077ece263feb90e3d58e6e237994 from
git@git.assembla.com:cavium/WindRiver.linux.git

The problem is if we have a running traffic on
the interface and as we are updating the sq->smq mapping on the fly,
we might loose completion interrupt for some packets. As a result of
that a watchdog reset is hit from BQL. To fix that issue we did
interface off/on during the process. But that is not a proper fix since
interface should not change its state during PFC updation.
This patch fixes the issue by halting transmission of the interface
during the PFC updation instead of full interface off. We enable the
transmission on the interface once PFC configuration is updated.

Fixes: 65914ddbf266 (octeontx2-pf: Update PFC configuration)
Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: I2bd5f5968273f3d6b6fde5e1337602c1793cf362
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/110861
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>