ptp: xilinx: Add workaround for EXTTS interrupt handling
commit 38b9e3f36bbcee6548139d653281f4114d7a7780 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

A possible IP or design related issue is observed where
multiple interrupt events are reported with a frequency
higher than expected (once per second). The timestamps in the
snapshot register are stale during this event as timestamp
is only captured on the rising edge of 1PPS. This issue only
occurs when EXTTS interrupt is enabled and disabled in
xlnx_ptp_enable function (this is typically called by testptp
or other PTP user space helpers to turn this feature ON/OFF)
As a workaround,
- Check whether EXTTS event is enabled before reporting event
timestamp.
- Don't disable/enable interrupts in xlnx_ptp_enable. Enable
them in probe and disable in remove.
There are no other interrupts enabled currently and this has no
side effects. The above ensures that stale timestamp is not
reported.

Signed-off-by: Harini Katakam <harini.katakam@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>