octeontx2-af: Move mutex lock to specific function
commit b3f329e452adc9a10c22adcf8232d1ca8e9f570b from
git@git.assembla.com:cavium/WindRiver.linux.git

Mutex lock is used to avoid race condition while
accessing the rvu_smmu_* registers. But current
code try to acquire the lock in wrapper function.
which is not needed. This patch moved to the internal
function in which these register are used.

fixes: f387ffd67ea8 (octeontx2-af: Secure APR table update with the lock)
Signed-off-by: Geetha sowjanya <gakula@marvell.com>
Change-Id: I1415d5c72441822a6126827eba7cac83cadc44b2
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/111888
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>