gpio: s32cc: Apply W/A for generic_handle_irq with RT_FULL
commit d451a6ad4ca93ac6a74e463510aceb811a3473fa from
https://github.com/nxp-auto-linux/linux

According to Documentation/gpio/driver.txt, "W/A: The
generic_handle_irq() is expected to be called with IRQ disabled, so the
IRQ core will complain if it is called from an IRQ handler which is
forced to a thread. The "fake?" raw lock can be used to W/A this
problem".

Issue: ALB-10464
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>