i2c: designware: introduce a custom scl recovery for SoCFPGA platforms
commit e817761fadfc54afe66fd1c05f3f19be0e187b0c from
https://github.com/altera-opensource/linux-socfpga.git socfpga-6.6.22-lts.

[upstream commit b2f670af1feab3325cce5a251128f30711148771]

The I2C pins on the SoCFPGA platforms do not go through a GPIO module,
thus cannot be recovered by the default method of by doing a GPIO access.
Only a reset of the I2C IP block can a recovery be successful, so this
change effectively resets the I2C controller, NOT any attached clients.

Reviewed-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Dinh Nguyen <dinguyen@kernel.org>
Signed-off-by: Liwei Song <liwei.song@windriver.com>