net: ethernet: cadence: Convert to platform remove callback returning void
commit 1a357b06f4a6498a00b0857765243d9dcfd39e39 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

The .remove() callback for a platform driver returns an int which makes
many driver authors wrongly assume it's possible to do error handling by
returning an error code. However the value returned is ignored (apart
from emitting a warning) and this typically results in resource leaks.
To improve here there is a quest to make the remove callback return
void. In the first step of this quest all drivers are converted to
.remove_new() which already returns void. Eventually after all drivers
are converted, .remove_new() is renamed to .remove().

Trivially convert this driver from always returning zero in the remove
callback to the void returning variant.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Reviewed-by: Claudiu Beznea <claudiu.beznea@tuxon.dev>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: harish h <harish.h@windriver.com>