spi: spi-cadence: Correct irq sequence for read
commit b9d25b16eb5bb405aa933406e211b521f09ec9ff from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

In current implementation during a read operation the driver reads one
extra byte than the number of bytes requested by spi-nor(from  RXFIFO
to system memory). In this process it tries a modify a memory location
that isn't allocated for the read operation. If the memory that driver
was trying to write is an read-only memory then this results in kernel
crash. Updated the driver irq sequence to return immediately after
finalizing the transfer.

Fixes: 46acc8a4ccbe ("spi: spi-cadence: Fix transfer timeout issue")
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>