pmdomain: arm: Add the SCMI performance domain
commit 687dab02e6db19b48849ec7e0f2ee3ae8219a524 from
https://github.com/nxp-imx/linux-imx.git

To enable support for performance scaling (DVFS) for generic devices with
the SCMI performance protocol, let's add an SCMI performance domain. This
is being modelled as a genpd provider, with support for performance scaling
through genpd's ->set_performance_state() callback.

Note that, this adds the initial support that allows consumer drivers for
attached devices, to vote for a new performance state via calling the
dev_pm_genpd_set_performance_state(). However, this should be avoided as
it's in most cases preferred to use the OPP library to vote for a new OPP
instead. The support using the OPP library isn't part of this change, but
needs to be implemented from subsequent changes.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Link: https://lore.kernel.org/r/20230919121605.7304-1-ulf.hansson@linaro.org
 [Move perf domain driver to firmware]
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>