spi: fsl-dspi: add cs-gpios support
commit 86eb4af6ec47c4d249c7dd915c380ad323f1ec3f from
https://github.com/nxp-auto-linux/linux

Make the driver be able to bit-bang a GPIO for the Chip Select pin of
select peripherals.

The GPIO value is driven by the driver in that case, and none of the
hardware Chip Select bits will be populated in the PUSHR register for
the TX commands constructed for this peripheral.

Issue: ALB-10584

Signed-off-by: Radu Pirea (NXP OSS) <radu-nicolae.pirea@oss.nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Link: https://lore.kernel.org/r/20221111211356.545026-1-vladimir.oltean@nxp.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 84b60f2bce1f0416617c4dbc015b5da906b8e2ad)
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>