arm64: dts: ti: k3-j721s2: Add NAVSS SRAM node for Codec
commit 8baa7229391868d4fe091d3d13ccf5f062d3f783 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Codec was using 128KB in 4MB of MSMC SRAM. But it is being mapped completely
for TI deeplearning usecases. So, adding the node for available 64KB NAVSS SRAM and
use it for Codec. Codec uses it for storing meta data to improve memory
bandwidth.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Reviewed-by: Rahul T R <r-ravikumar@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>