arm64: dts: ti: k3-j784s4: Add NAVSS SRAM node for Codec
commit 0c6d304b41c1f528fb64a21d65a4ed29a36dc65f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Codec was using 128KB in 8MB of MSMC SRAM. But it is being mapped completely
for TI Deeplearning usecases. So, adding the node for available 64KB NAVSS SRAM and use
it for codec. Codec uses it for storing meta data to improve memory
bandwidth.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Reviewed-by: Rahul T R <r-ravikumar@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>