spi: omap2-mcspi: Fix hardcoded reference clock
commit b07a95cd1bf15317c21e02000b6946f242844482 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 2d9f4877988f64f0f336983de65c365b6a7debfb upstream.

A hardcoded reference clock of 48 MHz is used to calculate the
clock divisor values, but the reference clock frequency can be
different across devices and can be configured which can cause
a mismatch between the reported frequency and actual SPI clock
frequency observed. Fix this by fetching the clock rate from
the clock provider and falling back to hardcoded reference only
if the clock is not supplied.

Fixes: 2cd7d393f461 ("arm64: dts: ti: k3-am654: Add McSPI DT nodes")

Reviewed-by: Dhruva Gole <d-gole@ti.com>
Link: https://lore.kernel.org/r/20230926113812.30692-1-vaishnav.a@ti.com
Signed-off-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Vaishnav Achath <vaishnav.a@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>