LF-10074: arm64: dts: imx8dxl-evk: correct the ext PHY reset sequence
commit dde3fb9802f2cc97165ffc88c8b40024109a3d83 from
https://github.com/nxp-imx/linux-imx.git

The external I2C GPIO controller and the external PHY both use the same
GPIO for reset purposes. However, the PHY requires that the I2C GPIO
controller is reset and initialized first. Therefore, the reset GPIO
should be moved out of the PHY initialization sequence. Instead, it should
be reset when the I2C GPIO controller is probed.

Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Reviewed-by: Frank Li <frank.li@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>