net: explicitly check in of_phy_is_fixed_link() for managed = "in-band-status"
commit  9694e62e0c74da09a6312571a21ba4b1c6a3ff0d from
https://github.com/nxp-imx/linux-imx lf-6.1.y

For a link described in OF using managed = "in-band-status",
of_phy_is_fixed_link() will return true for whatever reason.

We want to preserve that behavior, but we want to introduce another
value for the "managed" property, and we don't want
of_phy_is_fixed_link() to return true for it. So, we need to be explicit
about the check.

Documentation/devicetree/bindings/net/ethernet-controller.yaml makes it
clear that the 2 allowed values for "managed" are "auto" and
"in-band-status".

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>