arm64: dts: ti: k3-j784s4-main: Add the main domain watchdog instances
commit 20af8c446ba60f1b19a3680d68a8c1d0f1523545 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit caae599de8c6cc7405dcfd3bcd66a467eee23d2b upstream.

There are totally 19 instances of watchdog module. One each for the
8 A72 cores, one each for the 4 C7x cores, 1 for the GPU, 1 each
for the 6 R5F cores in the main domain. The non-A72 instances are
coupled with the R5Fs, C7x & GPU instances. Keeping them as reserved as
they are not used by A72.

Signed-off-by: Keerthy <j-keerthy@ti.com>
Link: https://lore.kernel.org/r/20231008044657.25788-5-j-keerthy@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>