clocking-wizard: Add support for versal clocking wizard
commit 3c4e86d571585620e1e5cec675e8cca952455324 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Add support for Clocking Wizard for Versal adaptive compute
acceleration platforms. The Versal clocking wizard differs
in the programming model and the register layout.
The CLKFBOUT_1 registers are at offset of 0x200
instead of the 0x330 in Versal. In Versal clocking wizard the low and
high time is programmed instead of the divisor.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
Link: https://lore.kernel.org/r/20231009071712.26469-3-shubhrajyoti.datta@amd.com
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>