spi: spi-fsl-lpspi: Add "cs_change" support
commit b6e776b317547c0fa0bec8c8d047981fbcf8c981 from
https://github.com/nxp-auto-linux/linux

According to the documentation, SR.FCF is set when frame is complete, and
PCS has negated. If cs_change is set and the transfer is the last one in
the message, PCS will remain asserted. In this case, SR.WCF will be set
instead of SR.FCF. The interrupt logic was changed to handle this case
also.

Issue: ALB-10628
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>