net: mdio-mux: be compatible with parent buses which only support C45
commit 07fe5814c38bad00f453dfe9fbf1c09b0401b1c5 from
https://github.com/nxp-imx/linux-imx.git

After the mii_bus API conversion to a split read() / read_c45(), there
might be MDIO parent buses which only populate the read_c45() and
write_c45() function pointers but not the C22 variants.

We haven't seen these in the wild paired with MDIO multiplexers, but
Andrew points out we should treat the corner case.

Link: https://lore.kernel.org/netdev/4ccd7dc9-b611-48aa-865f-68d3a1327ce8@lunn.ch/
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>