net: pcs: lynx: use MTIP AN/LT block for copper backplanes
commit de750c69a7ec7c4e997aa7d0182ec4d0dc217b1a from
https://github.com/nxp-imx/linux-imx.git

If the phylink_pcs is in the MLO_AN_C73 mode, then the Lynx PCS makes
use of the AN/LT block to advertise the supported backplane link modes
using clause 73 autoneg.

Note that we could also be advertising BASE-CR link modes (for SFP28
modules) but we don't have a way of detecting the medium type, so we
just hardcode backplane (BASE-K) for now.

Note that we find out whether we operate in MLO_AN_C73 mode a bit late
(later than lynx_pcs_create()). So we need mtip_backplane_create() to
not actually do anything until we know that C73 is required, and delay
any configuration until the phylink state machine kicks in.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>