thermal/drivers/k3_j72xx_bandgap: Use bool for i2128 erratum flag
commit 83b6b872cda56104a2050124b95b91500810d782 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 311f328ffc7572219bee65db77645e5fedd4e8e6 upstream

Some of TI's J721E SoCs require a software trimming method to report
temperatures accurately. Currently we are using a few different data
types to indicate when we should apply the erratum.

Change the 'workaround_needed' variable's data type to a bool to align
with how we are using this variable currently.

Signed-off-by: Bryan Brattlof <bb@ti.com>
Link: https://lore.kernel.org/r/20221031232702.10339-3-bb@ti.com
Signed-off-by: Daniel Lezcano <daniel.lezcano@kernel.org>
Signed-off-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>