dt-bindings: soc: microchip: add a property for system controller flash
commit 2fb13cd0e552ee95fbec07d71071f28228493a41 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

The system controller "shares" a SPI flash device with a QSPI controller
in the MSS. This flash is used to store FPGA bitstreams & other
metadata. IAP and Auto Upgrade both write images to this flash that the
System Controller will use to re-program the FPGA.

Add a phandle property signifying which flash device is connected to the
system controller.

Reviewed-by: Rob Herring <robh@kernel.org>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>