soc: microchip: mpfs: enable access to the system controller's flash
commit `059c2ab5493d9b5cac12c63b67079b264ce515a0 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

The system controller has a flash that contains images used to reprogram
the FPGA using IAP (In-Application Programming).
Introduce a function that allows a driver with a reference to the system
controller to get one to a flash device attached to it.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>