dt-bindings: gpu: ti: rogue: add am62p node updates
commit 83bdbf47e958e9cfa81e1bf2923e1ef0c77fcfdb from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The driver was updated for AM62P to include a new optional property:
assigned-clock-rates. This property can be used to override the driver's
default frequency of 800MHz for the AM62P variant. This will be rolling
out to other devices soon.

In addition to this the second power domain had to be associated with
this device. Updating power-domain-names for expected values. Also add
the clock name for the main clock for clock rate tracking in the driver.

Add an example of the AM62P device node.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>