LF-10523-17 firmware: arm_scmi: clock: support clock denied flags
commit 9f57854bc98af4159bc195d0404e352cb4e7ff35 from
https://github.com/nxp-imx/linux-imx.git

The firmware may export flags to indicate whether the clock
is allowed to set rate, set parent, enable/disable from the Agent.

If Agent is not allowed to enable/disable, directly return success.
If Agent is not allowed to set rate/parent, directly return -EACCES to
avoid SCMI RPC calls.

Signed-off-by: Peng Fan <peng.fan@nxp.com>
Acked-by: Nitin Garg <nitin.garg_3@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>