phy: ti: gmii-sel: Allow parent to not be syscon node
commit cd602d5975c892c74f98653d75ab168fd722bbbe from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 1fdfa7cccd35519d97a1f301b8143ea4196b2d16 upstream.

If the parent node is not a syscon type, then fallback and check
if we can get a regmap from our own node. This no longer forces
us to make the parent of this node a syscon node when that might
not be appropriate.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>