phy: ti: gmii-sel: Fix register offset when parent is not a syscon node
commit 2d2d7499e807fc30d902358325401e7b0a909cf7 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When the node for this phy selector is a child node of a syscon node then the
property 'reg' is used as an offset into the parent regmap. When the node
is standalone and gets its own regmap this offset is pre-applied. So we need
to track which method was used to get the regmap and not apply the offset
in the standalone case.

Fixes: 1fdfa7cccd35 ("phy: ti: gmii-sel: Allow parent to not be syscon node")
Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>