clk: microchip: mpfs: add missing MSSPLL outputs
commit a088c9ecc9dc8ec0a1ddc9fe334702bdaa0c2f68 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

The MSSPLL has 4 outputs, of which only the cpu/axi/ahb clock parent is
currently implemented.
Add the CAN clock too, as that'll be needed by the driver for the CAN
controller and uses output 3.
While we are here, the other two missing clocks, used by the eMMC/SD
controller and by the "user crypto".

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>