arm64: dts: ti: Add overlay for enabling RPi header for AM625-SK
commit 9fc2b558865670d7832b7e8d0f009214d84f3dd8 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

AM625-SK has GPIO and eHRPWM signals routed to the 40 pin header (J3).
This overlay will set the appropriate pinmux and enable GPIO and
PWM on those pins.

PWM0 A and B are routed to pin 12 and 40 respectively.
PWM1 A and B are routed to pin 36 and 33 respectively.

The gpio-hog p25 is needed because some pins are behind SoC UART1
FET SWITCH which is default HIGH and needs to be unset to route PWM0
A and B signals to the user expansion header.

The gpio-hog p05 and p06 are need to enable 3.3V and 5V rails on the
user expansion header.

Reviewed-by: Bryan Brattlof <bb@ti.com>
Signed-off-by: Judith Mendez <jm@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>