arm64: dts: ti: Add overlay for enabling RPi header for AM62A-SK
commit 121dcd3f649c790a1f265736690fbe4db62d6b32 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

AM62A-SK has GPIO and eHRPWM signals routed to the 40 pin header (J3).
This overlay will set the appropriate pinmux and enable GPIO and
PWM on those pins.

PWM0 A and B are routed to pin 35 and 12 respectively.
PWM1 A and B are routed to pin 36 and 33 respectively.

The gpio-hog p25 is needed because some pins are behind SoC UART1
FET SWITCH which is default HIGH and needs to be unset to route PWM0
A and B signals to the user expansion header.

The gpio-hog p05 and p06 are need to enable 3.3V and 5V rails on the
user expansion header.

Reviewed-by: Bryan Brattlof <bb@ti.com>
Signed-off-by: Judith Mendez <jm@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>