arm64: media: wave5: Add Support for frequency points
commit 1bb45095919cd3307dd7729168754a3212b7a119 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Support for OPP table is added to the driver. Driver can now select the
frequency from the available operational frequencies from the OPP table and
run at the frequency that is required for the Pixel rate that the IP is
handling. The Operating frequency adjusts to higher values when multiple
instances are initialized asynchronously and vice-versa. When no
instance is running, the frequency is set to the lowest operational
frequency of the OPP table.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Acked-by: Brandon Brnich <b-brnich@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>