drivers/net: dwmac-s32cc: Add fallback mechanism for accessing GPRs
commit 83b7a7c0e2812241205a4173c695ea4be6db50d3 from
https://github.com/nxp-auto-linux/linux

If the NVMEM provider required by the GMAC driver is not available
(e.g.: in Xen passthrough use-cases), we should resort to directly
accessing the required GPRs for PHY Interface selection mode
during the probe of the GMAC driver.

Added the required register back as a resource in the gmac0 DT
node, which can be mapped by the driver and used if NVMEM unavailable.

Issue: ALB-10755
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>