usb: gadget: uvc_video: unlock before submitting a request to ep
commit 57dde79828df828fa04bad470cd4b9c5a1578a9a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

There could be chances where the usb_ep_queue() could fail and trigger
complete() handler with error status. In this case, if usb_ep_queue()
is called with lock held and the triggered complete() handler is waiting
for the same lock to be cleared could result in a deadlock situation and
could result in system hang, so call usb_ep_queue() without the lock to
resolve this issue.

Signed-off-by: Piyush Mehta <piyush.mehta@amd.com>
Link: https://lore.kernel.org/r/20231108145426.2358406-1-piyush.mehta@amd.com
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>