drm/tidss: Fix dss reset
commit 8a74f75228b7d533d86a59a5f6197c5144ab1d16 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The probe function calls dispc_softreset() before runtime PM is enabled
and without enabling any of the DSS clocks. This happens to work by
luck, and we need to make sure the DSS HW is active and the fclk is
enabled.

To fix the above, and in preparation of adding boot time splash-screen
support, add a new function, dispc_init_hw(), which does:

- pm_runtime_set_active()
- clk_prepare_enable(fclk)
- dispc_softreset().

This ensures that the reset can be successfully accomplished and gives as
a place to add splash-screen code in the following patches.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>