media: v4l2: Add ignore_streaming flag
commit be98db2b3bb3c064cd264936015e028b1b0972d9 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 103d15bc20086820ce2c2af40a651529a87c9e3c upstream.

Add a new flag to the `struct v4l2_m2m_dev` to toggle whether a queue
must be streaming in order to allow queuing jobs to the ready queue.
Currently, both queues (CAPTURE & OUTPUT) must be streaming in order to
allow adding new jobs. This behavior limits the usability of M2M for
some drivers, as these have to be able, to perform analysis of the
sequence to ensure, that userspace prepares the CAPTURE queue correctly.

Signed-off-by: Sebastian Fricke <sebastian.fricke@collabora.com>
Signed-off-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>
Link: https://lore.kernel.org/r/20231108-wave5-v14-rebased-v14-1-0b4af1258656@collabora.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>