dt-bindings: counter: Update binding for K3 devices
commit f2da4a8daecd810fc01701283c5e3c4bee4fcc6c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This binding is for the eQEP Module found in various K3 devices
including AM62x and AM64x platform.

The interrupts property was set to "sysclkout" in the binding with the
intention of introducing interrupts to the driver in the future. Since we
are now introducing interrrupts in the driver, update the property to
"fck" since k3 devices use this clock in device tree.

Update the example found in binding.

Signed-off-by: Judith Mendez <jm@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>