dt-bindings: usb/ti,am62-usb.yaml: Add vddcore supply
commit c3aec8f5f52e583e018a9878cbd858939135736e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The USB driver needs to configure a bit in the USB PHY control
register depending on the PHY core voltage used in the system.

The PHY core supply is the same as the vdd_core supply on the SoC.

Add dt-bindings for this.

Signed-off-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>