remoteproc: k3-r5: Introduce MCU wakeup handlers
commit 06a089e482c5b55cbf5129492e6e4ec0269bf612 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Introduce MCU wakeup handler support in the remote proc driver. When
MCU wakeup is enabled on a SUSPEND request, the driver sends  a SUSPEND
mailbox message to the remote core and waits for the ACK from the remote
core. The remote device is not stopped in this scenario.

On the other hand for a graceful shutdown the remote core is stopped
and a ti-sci request is to sent to the DM to turn off the LPSC.

Signed-off-by: Hari Nagalla <hnagalla@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>