firmware: ti_sci: Partial-IO support
commit 241d614821ca8e89f972dbc95047dcabc7f82130 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add support for Partial-IO poweroff. In Partial-IO a few modules can
generate a system wakeup.

The modules that can be wakeup sources are defined by a DT property. If
one of these has wakeup enabled, instead of a normal poweroff we are
entering the Partial IO mode to be able to wakeup on activity.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>