drivers: pinctrl-scmi: Initialize `pcf` for each pin before usage
commit faec003c41e0e0dd829dab23889d795214e50346 from
https://github.com/nxp-auto-linux/linux

On probe, the pinctrl-scmi driver allocates each `pm_pin` and its
members, including `pcf`, but its fields are left uninitialized.

The `pcf` (pinconf configuration) field was then being used uninitialized
from the start, and whenever the driver was requested to modify the
configuration for a pin, the `pcf` data could be either garbage
(if ECC disabled), or already zeroized (if ECC enabled, due to scrubber).

Starting from garbage data in the existing pin configuration led
to various inconsistent crashes in the pinctrl-driver, depending
on the garbage data, since some data was interpreted as pointer
and dereferencing it led to aborts.

Initialized the `pcf` field of each `pm_pin` with `devm_kzalloc`
instead.

Issue: ALB-10738
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>