mailbox: llce: Allocate fifos in ascending order
commit 025a5c430cb44818df6742146343218762f46ff6 from
https://github.com/nxp-auto-linux/linux

This commit tries to mitigate the impact of the LLCE FW issue in which
the TX ACK frames are sent to the wrong LLCE CAN interface. This
scenario is reproduced when the fifo id allocated for a controller is
not the same with the controller id.
Allocating the fifo ids in ascending error tries to limitate the impact
of the firmware issue.

Issue: ALB-10677
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>