arm_scmi: smc: save flags only if spin lock was acquired
commit 49dc71dd740d826bbbf7b4d9af3c7a1a48fee22d from
https://github.com/nxp-auto-linux/linux

Previously, if we had concurrent requests, we could end up in a
situation were incorrect flags were restored which resulted in a
core running with interrupts disabled. Then, schedutil would try to
run and discover that interrupts were disabled and it would, in turn,
issue the bug message.

Issue: ALB-10746

Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>