net: hsr: Add support for MC filtering at the slave device
commit 8536abfc322d6cb91d9b21d245c292d16936f5c1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When MC (multicast) list is updated by the networking layer due to a
user command and as well as when allmulti flag is set, it needs to be
passed to the enslaved Ethernet devices. This patch allows this
to happen by implementing ndo_change_rx_flags() and ndo_set_rx_mode()
API calls that in turns pass it to the slave devices using
existing API calls.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Ravi Gunasekaran <r-gunasekaran@ti.com>
Reviewed-by: Wojciech Drewek <wojciech.drewek@intel.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>