input: edt-ft5x06: Include I2C details in names for the devices
commit f6743652ff5fb85c339fbae05a214a766cc646cb from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

[Commit 2a6c3115f4142e23ca10c984d7f65ac8fb901cc2 rpi-linux]

libinput uses the input device name alone. If you have two
identical input devices, then there is no way to differentiate
between them, and in the case of touchscreens that means no
way to associate them with the appropriate display device.

Add the I2C bus and address to the start of the input device
name so that the name is always unique within the system.

Patch Link:
https://github.com/raspberrypi/linux/commit/2a6c3115f4142e23ca10c984d7f65ac8fb901cc2

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Aradhya Bhatia <a-bhatia1@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>