LF-10725: media: imx8-isi: fix kernel dump issue when open/close the same device node
commit 8e10e4e7d593044f3cdabbd4af254cbc4ad7f071 from
https://github.com/nxp-imx/linux-imx.git

When the video device is streaming and open/close it in other thread,
it will trigger kernel dump. The root cause is that the thread will
close the video device and driver will do clean operation in release
callback, but actually, the ISI channel still be used by the first
thread, so when new frame received, it triggers new interrupt but
the callback is cleaned by the second thread, so kernel dump are
triggered.

Reproduce steps:
1. v4l2-ctl -d "/dev/video2" -v width=1280,height=800,pixelformat=YUYV --stream-mmap
2. v4l2-ctl --list-dev

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reported-by: Jared Hu <jared.hu@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>