firmware: xilinx: Expand feature check to support all PLM modules
commit a451db9cb7e18415e706c7ace2c46e27aacb260d from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

To support feature check for all modules, append the module id of the API
that is being checked to the feature check API so it could be routed to
the target module for processing.

There is no need to check compatible string because the board information
is taken via firmware interface.

Co-developed-by: Saeed Nowshadi <saeed.nowshadi@amd.com>
Signed-off-by: Saeed Nowshadi <saeed.nowshadi@amd.com>
Signed-off-by: Jay Buddhabhatti <jay.buddhabhatti@amd.com>
Link: https://lore.kernel.org/r/20231129112713.22718-3-jay.buddhabhatti@amd.com
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>