arm64: dts: ti: k3-am62p: add dll delay mapping for MMC1/MMC2
commit 5da2f861d7d06b97ea386294469e1644b18666fc from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add the input and output delay values for the available speed
modes for the MMC controller for MMC1/MMC2 for the am62p5 allowing
it to operate that the highest speed modes available.

SDR104 speed mode has not been finalized yet, but the process is
to save the base value, then update if characterization says
otherwise.

Signed-off-by: Judith Mendez <jm@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>