arm64: dts: ti: k3-am65: Add AM652 DTSI file
commit 2c4a8e340365aa5aefc2d158855d8276da98b499 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit fcb97d190c3ca411b37d8dd0b2650fa1378f08d4 upstream.

The AM652 is basically a AM654 but with 2 cores instead of 4. Add
a DTSI file for AM652 matching AM654 except this core difference.

This removes the need to remove the extra cores from AM654 manually
in DT files for boards that use the AM652 variant. Do that for
the IOT2050 boards here.

Signed-off-by: Andrew Davis <afd@ti.com>
Reviewed-by: Bryan Brattlof <bb@ti.com>
Link: https://lore.kernel.org/r/20231205162358.23904-1-afd@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>