LF-10688: ARM: dts: ls1021a-twr: disable USB2 by default
commit 78e83394e135bf610bb8fd2c566d9de99347f850 from
https://github.com/nxp-imx/linux-imx.git

The TWR-ls1021a board doesn't have the USB2 port natively.  However the
USB2 pins are routed to the ELEV expansion ports and only available with
TWR-SER2 board connected.  So disable the USB2 by default and users
of TWR-SER2 board should update the dts to enable it or apply an
additional overlay.

Signed-off-by: Li Yang <leoyang.li@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>