IMXBSP-101 net: stmmac: fix wrong place to call pinctrl state change
commit a167e7b48fbf1ad28ddd853795747d6dd1a33a47 from
https://github.com/nxp-imx/linux-imx.git

The previous patch put the phylink_suspend after the
pm_runtime_force_suspend to fix an issue. But in phylink_suspend, it
will call mdio read/write. If the iomux of the pins of MDIO bus have
been changed to others, the MDIO bus will not functional.
So put pinctrl_pm_select_sleep_state to the right place to fix this issue.

Fixes: 90702dcd19c0 ("net: stmmac: fix MAC not working when system resume back with WoL active")
Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>