ptp: Add Xilinx PTP timer driver with EXTTS support
commit 4fdd3c5e861edb27103fa3f6942223749f1df56b from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Add Xilinx PTP timer driver to be used with Xilinx MRMAC Ethernet
driver. This IP provides a PTP timer with support for time and fine
adjustment. It consists of a system timer and port timer (which is
used by the Ethernet MAC for timestamping). The port timer is used
for all time reporting and adjustment.
Xilinx timer-syncer supports 1PPS input from an external device.
A timer snapshot is provided (via registers) at the time of the 1PPS
HW event and an interrupt is generated. Reads the snapshot of timer and
reports to the application when 1PPS even is enabled.
System period offset registers have changed in recent timer-syncer IP
and there is no way to distinguish old system period offsets(0x130
and 0x134) and new offsets(0x34 and 0x38).So use
xlnx,timer-syncer-1588-3.0 compatible string for new offsets.

Note: A possible IP or design related issue is observed where
multiple interrupt events are reported with a frequency
higher than expected (once per second). The timestamps in the
snapshot register are stale during this event as timestamp
is only captured on the rising edge of 1PPS. This issue only
occurs when EXTTS interrupt is enabled and disabled in
xlnx_ptp_enable function (this is typically called by testptp
or other PTP user space helpers to turn this feature ON/OFF)
As a workaround,
- Check whether EXTTS event is enabled before reporting event
timestamp.
- Don't disable/enable interrupts in xlnx_ptp_enable. Enable
them in probe and disable in remove.
There are no other interrupts enabled currently and this has no
side effects. The above ensures that stale timestamp is not
reported.

Signed-off-by: Harini Katakam <harini.katakam@xilinx.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Suraj Gupta <suraj.gupta2@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>