net: xilinx: axiethernet: Add USXGMII support
commit 655287ac31aacb4a21a33504e3a9fb381d6c4e4b from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Add support for USXGMII IP in axiethernet driver.
This IP has a MAC similar to 10G/25G IP and supports USXGMII phy protocol.
USXGMII phy supports speeds from 10Mbps to 10Gbps. Only one phy speed
can be advertised at a time and this choice is obtained from the user
via a devicetree property. Tested at 1G and 2.5G speeds.
Since the MAC functionality is similar to 10G/25G, the same mac type is
used in the config structure. USXGMII IP requires a GT reset and this is
added to axienet_device_reset with a check based on mac type.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Acked-by: Harini Katakam <harini.katakam@amd.com>
Signed-off-by: Suraj Gupta <suraj.gupta2@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>