net: xilinx: Fix phylink_of_phy_connect error on RGMII designs
commit 73cf1aff5ad943ed66768a531173d945417f7132 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

For PCS/PMA based designs where PHY mode can be sgmii or 1000BaseX,
the existing implementation checks for pcs-handle and phy-handle
properties. For other modes (e.g RGMII), the implementation does
not check for these properties which results in PHY failures.

Update the code to ensure that pcs-handle and phy-handle properties
are parsed for all cases except XXV and MRMAC which do not have PHY
in the designs.

Signed-off-by: Srinivas Neeli <srinivas.neeli@amd.com>
Reviewed-by: Harini Katakam <harini.katakam@amd.com>
Signed-off-by: Suraj Gupta <suraj.gupta2@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>