LF-10913 net: stmmac: dwmac-imx: add clock input support in RMII mode
commit fbc17f6f7919d03c275fc48b0400c212475b60ec from
https://github.com/nxp-imx/linux-imx.git

In RMII mode, the default setting is using the 50MHz reference clock
from SoC. The ref_clk pin is output.
If users want to use external or PHY to provide the 50MHz ref clock,
it needs to config the GPR register.
So, add the code to config GPR register to support the clock input.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>