LF-10932-1 net: enetc: add setting VF mac address for imx95
commit 7c534e64c0fe2d21e7f6bcd4cba3ef68f10b8ef9 from
https://github.com/nxp-imx/linux-imx.git

PF mac address will be provided by uboot or fuse.
The mac address of each VFn is obtained by adding n to the mac address
of the PF where it is located.(n = 1 or 2 in imx95)
Merge the call of set_si_primary_mac() into enetc_set_si_hw_addr().

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>