LF-10932-2 nvmem: imx95: add the support to return the mac addr of PFs
commit acc1d905aa0fe8fcaf9ba6993c24f574e8c7f527 from
https://github.com/nxp-imx/linux-imx.git

The 3 PFs of imx95 use 9 mac addresses, but only saves one mac base
address in fuse, which is the mac address of the first PF. The others
PF mac addresses need to be calculated through offsets.
According to the ocotp framework, adding the PF offset to bits 15~12 of
offset. This will not affect the dts node setting of other fuse devices.
In imx95, the value range of offset is 0~0x984.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Alice Guo <alice.guo@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>