LF-10972: mxc: vpu: wave6: don't clear drain flag in handling source change
commit a00dc5924e812030494573b69a0538b5396fc45d from
https://github.com/nxp-imx/linux-imx.git

Driver will try to initiate a drain of capture queue in dynamic
resolution change, but if user has started drain before that, driver may
clear the drain flag in handling source change, and it may led to
pipeline hang.

We should not clear the drain flag and set eos in handling source
change.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: TaoJiang <tao.jiang_2@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>