ARM: dts: bcm2711-rpi-400: Don't delete the ACT LED
commit a7479ae20b3725616e345e8c6423f5d56d89624b from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The Pi 400 only has one user-accessible LED (green), which is intended
to be used as a power indication. Therefore there is no activity LED.
However, a user may wish to remap the activity LED functionality to a
GPIO in the 40-way header, so preserve the led-act node in a disabled
state.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>