spi: spi-cadence-quadspi: Add support for odd byte count read/write in DDR mode
commit 8d52b0bd7ba8c65569db06b33186791f0f45fc75 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Due to the nature of DDR protocol, an even number of bytes is always
transferred and also the least significant bit of the byte address shall
always be zero when using DDR protocol.
Add support for read and write operations for the requests with odd byte
count or odd address or both. These kind of requests are expected with
UBIFS file system and mtd_stresstest.

Signed-off-by: Sai Krishna Potthuri <lakshmi.sai.krishna.potthuri@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>