spi: spi-cadence-quadspi: Add module param to configure read timeout
commit 6a768a2c9f12ebb632b463b1f4f51dc48c8d2ca0 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

On emulation platform, OSPI is operating at very low frequency and
due to various factors default 10msec defined in the driver is not
optimal and may lead to timeout because we are waiting on an interrupt.

Add module parameter support to configure the read timeout in msec.
For example to set 50msec timeout, this can be passed as a module
parameter to the builtin driver as shown below.
bootargs = "earlycon=pl011,mmio32,0xf1920000 console=ttyAMA0,115200
            spi-cadence-quadspi.read_timeout_ms=50";

Signed-off-by: Sai Krishna Potthuri <sai.krishna.potthuri@amd.com>
Signed-off-by: Michal Simek <michal.simek@amd.com>
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>