firmware: xilinx: Remove clock_setrate and clock_getrate api
commit 7508d2141c629a2774a2b4b0b968a1bfe5fe9f7b from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

As per the current code base, PM_CLOCK_SETRATE and PM_CLOCK_GETRATE
APIs are not supported for the runtime operations. In the case of
ZynqMP returning an error from TF-A when there is any request to
access these APIs and for Versal also it is returning an error like
NO_ACCESS from the firmware. So, just removing the unused code to
avoid the confusion around these APIs.

Also, there is no issue with the backward compatibility as these APIs
were never used since implemented. Hence no need to bump up the
version of the feature check API as well.

Signed-off-by: Ronak Jain <ronak.jain@amd.com>
Link: https://lore.kernel.org/r/6ccbffbafd1f0f48f6574d5a3bf2db6a5603fdb0.1702565618.git.michal.simek@amd.com
Signed-off-by: Michal Simek <michal.simek@amd.com>
State: waiting
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>