LF-10988 net: fec: add lock protection for pps_enable
commit 3077bbca8330cb8c893499814132bcd3b47d326f from
https://github.com/nxp-imx/linux-imx.git

The update of fep->pps_enable is protected by tmreg_lock, so the read
operation of fep->pps_enable should be also protected by tmreg_lock.
Otherwise, evasion of holding a lock in this way can allow operations
to be interleaved or reordered at runtime, such that data races occur.

Fixes: 278d24047891 ("net: fec: ptp: Enable PPS output based on ptp clock")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>