selftests: forwarding: ethtool_mm: support devices with higher rx-min-frag-size
commit 8105b7d265a7795101bb99523730d9f40e2676de from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 2491d66ae66cdd761c38563e17940a707d1b5e91 upstream.

Some devices have errata due to which they cannot report ETH_ZLEN (60)
in the rx-min-frag-size. This was foreseen of course, and lldpad has
logic that when we request it to advertise addFragSize 0, it will round
it up to the lowest value that is _actually_ supported by the hardware.

The problem is that the selftest expects lldpad to report back to us the
same value as we requested.

Make the selftest smarter by figuring out on its own what is a
reasonable value to expect.

Cc: Shuah Khan <shuah@kernel.org>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Tested-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: Roger Quadros <rogerq@kernel.org>
Tested-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Link: https://lore.kernel.org/r/20231219105805.80617-2-rogerq@kernel.org
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>