pinctrl: s32cc: revert changes for slew rate in Mhz
commit 0968dfb19912b3015eaaa23e232b90fc50c9df9b from
https://github.com/nxp-auto-linux/linux

Slew rate was being parsed in Mhz from the DT as part of
the initial s32g2 support in upstream (commit fd84aaa8173d).
However, this cannot be used at the moment since it does
not take into account the voltage level.
This can be used as a temporary fix until perhaps we decide
on a re-mapping between the values in DT bindings and the driver.

Issue: ALB-10855

Signed-off-by: Florin Buica <florin.buica@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>