gpio: siul2: configure the base
commit e7f088ac69044656122498194a1b55eb514ef4f9 from
https://github.com/nxp-auto-linux/linux

Configure the base for the GPIO driver. Even though it is deprecated to
set the base for the GPIO driver, the gpio pin ranges would not match
the pinctrl ranges if the base is not correctly configured. This means
that calls to retrieve the pinctrl driver associated to a GPIO pin might
fail.

Issue: ALB-10947

Signed-off-by: Andrei Stefanescu <andrei.stefanescu@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>