octeontx2-af: map management port always to first PF
commit 4667900e0f4e4e8a37dc8f0d7c32ad2ee97547de from
git@git.assembla.com:cavium/WindRiver.linux.git

The user can enable or disable any MAC block or a few ports of the
block. The management port's interface name varies depending on the
setup of the user if its not mapped to the first pf.

By changing the management port mapping to always point to the first PF,
this patch resolves the problem.

Change-Id: I0deec112a41f884c5d83d2a669d22a7b5a84974b
Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/119778
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>