mux: mmio: use reg property when parent device is not a syscon
commit b2a05120339a85852dbd6acd2bfe00633001a94d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 2765149273f4b52beb07256c3e4686c065a5f8a8 upstream.

The DT binding for the reg-mux compatible states it can be used when the
"parent device of mux controller is not syscon device". It also allows
for a reg property. When the reg property is provided, use that to
identify the address space for this mux. If not provided fallback to
using the parent device as a regmap provider.

While here use dev_err_probe() in the error path to prevent printing
a message on probe defer which now can happen in extra ways.

Signed-off-by: Andrew Davis <afd@ti.com>
Reviewed-by: Nishanth Menon <nm@ti.com>
Acked-by: Peter Rosin <peda@axentia.se>
Link: https://lore.kernel.org/r/20240104154552.17852-1-afd@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>