drm: xlnx: hdmi: Add support for AVI infoframes
commit 9dfb23da778bf6b0d4e6377fceaefa3f6a9abcfe from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Various aspects of the current video stream are indicated by the
HDMI source to the sink with an Auxiliary Video Information(AVI)
Infoframes. Source shall send the AVI to the sink once per video
field. The data from the AVI applies to the next full frame of the
video data. Refer section 8.2.1 in HDMI specification 1.4b.

YUV formats aren't working currently with the driver. As per the
spec (section 8.2.1 in HDMI specification 1.4b), AVI Infoframes
should be sent to fix this.

Signed-off-by: Rajesh Gugulothu <rajesh.gugulothu@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>