drm/vc4: Flush stale dlist entries if allocation fails
commit e33f46eba97edbaff7744bfd5c5a093cd2c567d3 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

This is largely for debug at present.
For reasons unknown we are not getting the end of frame interrupts
that should trigger a sweep of stale dlist entries.

On allocation failure clear out ALL stale entries, and retry the
allocation. Log the interrupt status so we have debug regarding
whether the HVS believes the interrupt is enabled.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>