LF-11363-1: arm64: dts: imx8mp-evk-basler: enable isi0 mem-to-mem feature
commit 2f6691783a461cf1b88b91e85fc38ffb12ac5e1b from
https://github.com/nxp-imx/linux-imx.git

Enable isi0 m2m for single Basler camera's dts file.
Because of the hardware design, just the isi0 can be used for m2m feature,
but the isi0 capture and m2m function can't be used at the same time. So
if enable the isi0 m2m function, the isi0 capture function should be
disabled.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>