LF-11363-3: arm64: imx8mp-evk-dual-os08a20: enable isi0 mem-to-mem feature
commit bf481545174732e4ba737eba89dd09a4fe310eef from
https://github.com/nxp-imx/linux-imx.git

Enable m2m in dual os08a20 cameras' dts file. Because of the hardware
design, just the isi0 can be used for m2m feature, but the isi0 capture
and m2m function can't be used at the same time. So if enable the isi0
m2m function, the isi0 capture function should be disabled.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>