LF-11319: arm64: dts: move VPU OCRAM to the top 96KB
commit 227e432c7877b6c2fcdc06a675cee0376e9f0a1b from
https://github.com/nxp-imx/linux-imx.git

The spl is stored in the ocram, and vpu use the same ocram, then the spl
may be overwritten by the vpu. It will make reboot hang.

Moving SPL to DDR would cause other potential issues, since SPL not use
all the OCRAM, so suggest move VPU OCRAM to the top 96KB which not
conflict with OCRAM.

The ocram range is [0x20480000, 0x204d8000), so change the vpu sram
to [0x204c0000, 0x204d8000)

Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: TaoJiang <tao.jiang_2@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>