riscv: dts: microchip: add resets to the MSS CAN
commit d583954e38edae5ca0457a420d264dca44114dc8 from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

The MSS CAN needs to be reset during the ned_dev operations like start,
stop and open to bring the device to the reset state. There is no separate
register in the MSS CAN hardware to do the soft reset hence add the reset
property to the CAN node to do the full device reset.

Signed-off-by: Naga Sureshkumar Relli <nagasuresh.relli@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>