arm64: dts: ti: k3-j7200: Enable PCIe nodes at the board level
commit 8ba0a7a6180383c290822928e7251955ba3de07e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 1b63a1b480c27764d30a0924a4982d31e15df6fd upstream.

PCIe node defined in the top-level J7200 SoC dtsi file is incomplete
and will not be functional unless it is extended with a SerDes PHY.

As the PHY and mode is only known at the board integration level, this
node should only be enabled when provided with this information.

Disable the PCIe node in the dtsi files and only enable when it is
actually pinned out on a given board.

Signed-off-by: Andrew Davis <afd@ti.com>
Link: https://lore.kernel.org/r/20240124183659.149119-1-afd@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>