arm64: dts: ti: k3-j7200: Remove PCIe endpoint node
commit 0b3b6de6a6cc4b230809ec1c3825ba627d449f4f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 0b16abe711bd5136f2319c4f548fb20588540266 upstream.

This node is an example node for the PCIe controller in "endpoint" mode.
By default the controller is in "root complex" mode and there is already a
DT node for the same.

Examples should go in the bindings or other documentation.

Remove this node.

Signed-off-by: Andrew Davis <afd@ti.com>
Link: https://lore.kernel.org/r/20240124183659.149119-2-afd@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>