arm64: dts: ti: k3-am65: Remove PCIe endpoint nodes
commit c5a592830ad729c612559427c0b572d5a6855d9e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit e074d9d9a52eec11cd8b3f98e2576b0c762686f1 upstream.

These nodes are example nodes for the PCIe controller in "endpoint" mode.
By default the controller is in "root complex" mode and there is already a
DT node for the same.

Examples should go in the bindings or other documentation.

Remove this node.

Signed-off-by: Andrew Davis <afd@ti.com>
Link: https://lore.kernel.org/r/20240124183659.149119-3-afd@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>