arm64: dts: ti: k3-am64: Remove PCIe endpoint node
commit 53dc2a96d1c22ca7c76da3ce8c708f4e103a5130 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 6cce60550763564360beadf57ecf2e4d45b585e1 upstream.

This node is an example node for the PCIe controller in "endpoint" mode.
By default the controller is in "root complex" mode and there is already a
DT node for the same.

Examples should go in the bindings or other documentation.

Remove this node.

Signed-off-by: Andrew Davis <afd@ti.com>
Link: https://lore.kernel.org/r/20240124183659.149119-4-afd@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>