arm64: zynqmp: Sync clock labels with kr260 revB
commit 0506e8f0b8cb20065e2240a3daedd2345985f1e2 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

Board description describes the hard part of chip (PS) but programmable
logic (PL) part is not described in this file. But clocks on the board are
not only connected to PS but also wired to PL. And because two revisions
are available where revA is using one si5332 and revB multiple clock chips
using the same clock labels helping with keeping only one device tree
overlay which targets PL. That's why synchronize clock labels and use
labels from revB which are more generic.
Unfortunately if there is driver for si5332 chip split could happen again
but it is still worth to do it now and solve this issue when occurs.

Reported-by: Sagar Karmarkar <sagar.karmarkar@amd.com>
Link: https://lore.kernel.org/r/abac6069e6029ed4076ec7b9d6b33604b6072aa3.1706253871.git.michal.simek@amd.com
Signed-off-by: Michal Simek <michal.simek@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>