thermal: qoriq: Add support for reading TMR.ALPF value from DTS
commit 568a803cc4d1f6c7f320479585e58407d05fffba from
https://github.com/nxp-auto-linux/linux

ALPF stands from "Average Low Pass Filter Setting"
The average temperature is calculated using this formula:
ALPF x Current_Temp + (1 - ALPF) x Average_Temp.
The following values should be written in TMR.ALPF:
00b => ALPF = 1.0
01b => ALPF =0.5
10b => ALPF =0.25
11b => ALPF = 0.125

If TMU version 2 is used, a new property can be added in the device
tree, "tmr-alpf", for replacing the default value which is 11b. This
implies that ALPF value used in the formula will be 0.125.

Issue: ALB-10839
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>