arm64: versal: Configure reset gpio as active low
commit fb145395f80f6c7ee8bb6ecc7a43a3a4998290e2 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

The device reset functionality of the spi-nor core expects the reset GPIO
to be active low, but it is currently configured as active high. As a
result, the spi-core fails to reset the OSPI flash device. Resolve this
issue by configuring the reset GPIO as active low in the device tree.

Fixes: 2e56fb683c9d ("arm64: versal: Move reset-gpios property from controller node to flash node")
Signed-off-by: Amit Kumar Mahapatra <amit.kumar-mahapatra@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>