ASoC: DACplusADCPro - fix 16bit sample support in clock consumer mode
commit 6b9407605037a3dfd06ec259a49a4d6c78ff24dd from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The former code did not adjust the physical sample width when in
clock consumer mode and has taken the fixed 32 bit default. This
has caused the audio to be played at half its frequency due to
the fixed bclk_ratio of 64.

Problem appears only on PI5 as on the former PIs the I2S module
did simply run at fixed 64x rate.

Signed-off-by: Joerg Schambacher <joerg@hifiberry.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>