Subject: [PATCH v6.1] powerpc: drop duplicate entry in Makefile for lib dir
We have two copies of "crtsavres: fixups for 5.4+" - presumably because
it is a commit with lots of whitespace in the context that will loosely
match in several places, leading to easily but inadvertently applying it
more than once.

In the v6.1/standard/base we have the following SHAs:

commit 77fa1b70cdbdf9c846592a8f7bdf4873b8cba9f9
commit cd107e4edd5f798689074cd08845db78e8834f13

The v5.10 and v5.15 did not have the duplicated v5.4 content, and the
v6.5 code base seems to have sensibly dropped the old fixup entirely.

The duplicated lines are obvious in the context of this fixup, and leads
to the following:

arch/powerpc/Makefile:383: warning: overriding recipe for target 'arch/powerpc/lib/crtsavres.o'

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>