drivers: staging: bcm2835-isp: Respect caller's stride value
commit 55faf2eb5190410941aaabdc5bbd2c74c27f546e from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The stride value reported for output image buffers should be at least
as large as any value that was passed in by the caller (subject to
correct alignment for the pixel format). If the value is zero (meaning
no value was passed), or is too small, the minimum acceptable value
will be substituted.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>