soc: xilinx: Remove the usage of get/put_cpu() in xlnx_event_init_sgi()
The stable commit 01946c3c83b2 ("soc: xilinx: Fix for call trace due to
the usage of smp_processor_id()") is trying to fix a call trace in the
xlnx_event_init_sgi(), but that issue has already been fixed by a local
commit 3eea40ee0991 ("driver: soc: xilinx: rename cpu_number1 to
dummy_cpu_number"). Compare to the changes in the stable commit, our
local commit is more reasonable. So revert the change of stable commit.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>