EDAC/versal: Report the bit position of the error
commit c39d7ed6002d4659fc2918e84dc5b107e676a685 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

In the case of CE, the bit position can be derived from the data
and the parity registers. The values are read and
then from the fixed H Matrix we are derive the bit
position that is corrected. The hardware already corrects the
CE errors however we are deriving it from the registers and
reporting it for debug purposes.

When a correctable error is detected, the driver reads from the
registers:

eccr0/eccr1_corr_err_data_lo
eccr0/eccr1_corr_err_data_hi
eccr0/eccr1_corr_err_data_par

and along with shifting/XORing fixed H_Matrix ECC values,
calculates which bit location is found, mimicking the calculation
done in the DDRMC ECC hw.

The H_Matrix was generated from an old 1950's IBM paper.

The maximum bits in the data read is 64 so the data match is run for
64 bits.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>