remoteproc: core: Make shutdown-on-release per-file handle
commit a4dac51a55b7950e8e3fdd71db15b8c1686a8a28 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

More than one file handle can be open for a single remote processor,
but the flag specifying that the core should be shutdown when a file
handle closes is global for a remote processor. This means that if
any file handle closes, the core is shutdown, even if that user
didn't set that on the current file handle.

Store an instance of the shutdown-on-release flag for each handle.
This allows for other users of the rproc cdev to use and close
file handles without accidentally shutting off the core.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>