HSD #15015532380: fpga: altera-cvp: Fixed credit register
commit cf1c418ce778ec6442ccd5b6500a69edb64bbc35 from
https://github.com/altera-opensource/linux-socfpga.git socfpga-6.6.22-lts.

This fix is required in order to get  Agilex5 SOC FPGA CVP bitstream
transfer done with the new credits changes.
Fixed the Agilex5 credits to be 12 bits instead of 8 bits in non-Agilex5
devices. Also added debug messages to make it clear to the user
that the currently attached device type family as well as status
of the CVP completion.

Signed-off-by: Ang Tien Sung <tien.sung.ang@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>