s32cc: pcie: initialize EP entries to 0
commit e9522b92c6ab47b4a13b412249c53ba010247c19 from
https://github.com/nxp-auto-linux/linux

Uninitialized data allocated for EP entries in the EP list may lead to
unexpected behaviours when configuring the EP BARs, e.g. the size of an
EP bar may not be 0 before starting the configuration. Use kzalloc()
instead of kmalloc() to initialize all fields of the EP entry to 0.

Issue: ALB-11147
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>