ARM: dts: bcm2712: Move /soc/sound to bcm2712-rpi
commit 82f410643cb9bdb18be919c06e9a7cba5f2a3783 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

The /soc/sound node is not a feature of the bcm2712 hardware, rather a
Raspberry Pi convention of where to declare soundcards. Move it to
bcm2712-rpi.dtsi accordingly, at the same time marking it as disabled
so that runtime overlay application causes the soundcard device to be
created.

See: https://forums.raspberrypi.com/viewtopic.php?t=367530

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>